+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Revamped the api surface for request configuration. [#1494](https://github.com/microsoft/kiota/issues/1494)
+
 ## [0.0.23] - 2022-04-19
 
 ### Changed